1-array_iterator.c

+++ 0x0F-function_pointers/function_pointers.h
@@ -4,6 +4,6 @@
 
 void print_name(char *name, void (*f)(char *));
 int _putchar(char c);
-void array_iterator(int *array, size_t size, void (*action)(int));
+void array_iterator(int *array, unsigned int size, void (*action)(int));
 
 #endif
